backend/requirements.txt
fastapi==0.109.2
langchain==0.2.7
langchain_community==0.2.7
langchain_groq==0.1.6
pydantic==2.7.2
python-dotenv==1.0.1
langchain_huggingface==0.0.3
faiss-cpu==1.8.0.post1
uvicorn==0.27.1